py/requirements.txt
asgiref==3.8.1
Brotli==1.0.9
brotlicffi==1.1.0.0
ConfigParser==7.2.0
contextlib2==21.6.0
cryptography==44.0.2
cx_Oracle==8.3.0
docutils==0.21.2
fastapi==0.115.12
h2==4.2.0
# HTMLParser==0.0.2
huggingface_hub==0.29.1
importlib_metadata==8.6.1
ipython==8.12.3
ipywidgets==8.1.5
jnius==1.1.0
js==1.0
keyring==25.6.0
langchain==0.3.24
langchain_community==0.3.22
langchain_core==0.3.56
langchain_huggingface==0.1.2
langchain_openai==0.3.14
lockfile==0.12.2
nacl==0.0.0
objgraph==3.6.2
openai==1.76.0
Pillow==11.2.1
protobuf==6.30.2
psutil==6.1.1
pydantic==2.11.3
pyodide==0.0.2
pyOpenSSL==25.0.0
pysqlcipher3==1.2.0
pytest==8.3.5
python-dotenv==1.1.0
python_jose==3.4.0
railroad==0.5.0
simplejson==3.20.1
Sphinx==8.2.3
sqlcipher3==0.5.4
thread==2.0.5
tornado==6.4.2
truststore==0.10.1
urllib3_secure_extra==0.1.0
uvicorn==0.34.2
watchdog==6.0.0
zstandard==0.23.0
cryptography>=3.4.8
